+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Add tests for ``.as_float()`` modifier usage with wiring.
+
 4.26.0
 ------
 - Add wiring by string id.